logError function
