+++ src/service/graphql/middlewares/error.ts
@@ -51,8 +51,7 @@ export async function error (ctx: GraphQLServiceContext, next: () => Promise<voi
     await next()
 
     graphQLErrors = parseErrorResponse(ctx.graphql.graphqlResponse || {})
-  }
-  catch (e) {
+  } catch (e) {
     const formatError = ctx.graphql.formatters!.formatError
 
     if (e.isGraphQLError) {
@@ -69,8 +68,7 @@ export async function error (ctx: GraphQLServiceContext, next: () => Promise<voi
     if (e.headers) {
       ctx.set(e.headers)
     }
-  }
-  finally {
+  } finally {
     if (graphQLErrors) {
       console.error('[node-vtex-api graphql errors]', graphQLErrors)
       ctx.graphql.status = 'error'
