+++ src/service/http/middlewares/error.ts
@@ -1,5 +1,6 @@
 import { IOClients } from '../../../clients/IOClients'
 import { cleanError } from '../../../utils/error'
+import { logError } from '../../../utils/error'
 import { ServiceContext } from '../../typings'
 
 const CACHE_CONTROL_HEADER = 'cache-control'
@@ -34,42 +35,6 @@ export async function error<T extends IOClients, U, V> (ctx: ServiceContext<T, U
       ctx.set(CACHE_CONTROL_HEADER, `no-cache, no-store`)
     }
 
-    // Log error
-    const {
-      method,
-      status,
-      vtex: {
-        operationId,
-        requestId,
-        route: {
-          id,
-        },
-      },
-      headers: {
-        'x-forwarded-path': forwardedPath,
-        'x-forwarded-host': forwardedHost,
-        'x-forwarded-proto': forwardedProto,
-        'x-vtex-platform': platform,
-      },
-    } = ctx
-
-    const log = {
-      ...err,
-      forwardedHost,
-      forwardedPath,
-      forwardedProto,
-      method,
-      operationId,
-      platform,
-      requestId,
-      routeId: id,
-      status,
-    }
-
-    // Use sendLog directly to avoid cleaning error twice.
-    ctx.clients.logger.sendLog('-', log, 'error').catch((reason) => {
-      console.error('Error logging error 🙄 retrying once...', reason ? reason.response : '')
-      ctx.clients.logger.sendLog('-', log, 'error').catch()
-    })
+    logError(ctx, err)
   }
 }
